+++ test/pow/phoenix/mailer/mail_test.exs
@@ -40,8 +40,8 @@ defmodule Pow.Phoenix.Mailer.MailTest do
     assert mail = Mail.new(conn, :user, {MailerView, :mail_test}, value: "test")
 
     assert mail.user == :user
-    assert mail.subject == ":web_mailer_module subject :user"
-    assert mail.html == "<p>:web_mailer_module html mail :user</p>"
-    assert mail.text == ":web_mailer_module text mail :user"
+    assert mail.subject == ":web_mailer_module subject :user :custom_assign"
+    assert mail.html == "<p>:web_mailer_module html mail :user :custom_assign</p>"
+    assert mail.text == ":web_mailer_module text mail :user :custom_assign"
   end
 end